Finalização projeto

+++ src/br/ucb/modelo/Editora.java
@@ -14,7 +14,7 @@ public class Editora implements Serializable {
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 	private String nome;
-	private Long votos;
+	private int votos;
 
 	public Editora() {
 
@@ -40,11 +40,11 @@ public class Editora implements Serializable {
 		this.nome = nome;
 	}
 
-	public Long getVotos() {
+	public int getVotos() {
 		return votos;
 	}
 
-	public void setVotos(Long votos) {
+	public void setVotos(int votos) {
 		this.votos = votos;
 	}
